refactor: Update folder structure for layouts
