+++ src/index.ts
@@ -7,5 +7,9 @@ export { Spinner, type SpinnerProps } from '@components/Spinner';
 export { Text, type TextProps } from '@components/Text';
 export { Tooltip, type TooltipProps } from '@components/Tooltip';
 
-export * from '@layouts';
+export { Box, type BoxProps } from '@layouts/Box';
+export { Container, type ContainerProps } from '@layouts/Container';
+export { Flex, type FlexProps } from '@layouts/Flex';
+export { Grid, type GridProps } from '@layouts/Grid';
+
 export * from '@utils';
