+++ src/layouts/Box/index.ts
@@ -1,0 +1,2 @@
+export { Box } from './Box';
+export type { BoxProps } from './Box';
