+++ src/layouts/Container/index.ts
@@ -1,0 +1,2 @@
+export { Container } from './Container';
+export type { ContainerProps } from './Container';
