+++ src/layouts/Flex/index.ts
@@ -1,0 +1,2 @@
+export { Flex } from './Flex';
+export type { FlexProps } from './Flex';
