+++ src/layouts/Grid/Grid.test.tsx
@@ -3,7 +3,7 @@ import { vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import Grid from './Grid';
+import { Grid } from './Grid';
 
 describe('[Layouts] Grid', () => {
   describe('rendering', () => {
