+++ src/layouts/Grid/index.ts
@@ -1,0 +1,2 @@
+export { Grid } from './Grid';
+export type { GridProps } from './Grid';
