--- src/layouts/index.ts
@@ -1,4 +1,0 @@
-export { default as Box } from './Box';
-export { default as Container } from './Container';
-export { default as Flex } from './Flex';
-export { default as Grid } from './Grid';
